게시글과 유저 연결(작성자 유저로)

+++ src/main/java/com/example/board/service/PostService.java
@@ -1,11 +1,16 @@
 package com.example.board.service;
 
 import com.example.board.domain.Post;
+import com.example.board.dto.PostDto;
 import com.example.board.exception.PostNotFoundException;
 import com.example.board.repository.PostRepository;
+import com.example.board.security.User;
+import com.example.board.security.details.Role;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,11 +20,6 @@ public class PostService {
 
     private final PostRepository postRepository;
 
-
-//    public List<Post> findAll() {
-//        return postRepository.findAll();
-//    }
-
     public Page<Post> findAll(Pageable pageable) {
         return postRepository.findAll(pageable);
     }
@@ -27,22 +27,36 @@ public class PostService {
         return postRepository.findById(id).orElseThrow(()->new PostNotFoundException("해당 게시글을 찾을 수 없습니다."));
     }
 
-    public Post create(Post post) {
+    public Post create(@Valid PostDto postdto, User user) {
+        Post post = new Post();
+        post.setUser(user);
+        post.setTitle(postdto.getTitle());
+        post.setContent(postdto.getContent());
         return postRepository.save(post);
     }
 
-    public void delete(Long id) {
-        postRepository.deleteById(id);
+    public void delete(Long id, User user) {
+        Post post = findById(id);
+
+        if(isWriterOrAdmin(user,post)){
+            postRepository.deleteById(id);
+        }
+        else {
+            throw new AccessDeniedException("게시글을 삭제할 권한이 없습니다.");
+        }
     }
 
 
     @Transactional
-    public void update(Long id, Post updatedPost) {
-        postRepository.findById(id).ifPresent(post -> {
-            post.setTitle(updatedPost.getTitle());
-            post.setContent(updatedPost.getContent());
-            post.setAuthor(updatedPost.getAuthor());
-        });
+    public void update(Long id, PostDto postDto, User user) {
+        Post post = postRepository.findById(id).orElseThrow(() -> new PostNotFoundException("게시물이 존재하지 않습니다."));
+
+        if(!isWriterOrAdmin(user,post)) {
+            throw new AccessDeniedException("게시글에 수정할 권한이 없습니다.");
+        }
+
+        post.setContent(postDto.getContent());
+        post.setTitle(postDto.getTitle());
 
     }
 
@@ -54,7 +68,7 @@ public class PostService {
             case "content":
                 return postRepository.findByContentContaining(keyword, pageable);
             case "author":
-                return postRepository.findByAuthorContaining(keyword, pageable);
+                return postRepository.findByUser_EmailContaining(keyword, pageable);
             default:
                 return postRepository.findAll(pageable);
         }
@@ -65,6 +79,9 @@ public class PostService {
         Post post = postRepository.findById(id).orElseThrow(() -> new PostNotFoundException("게시글이 존재하지 않습니다."));
         post.setViews(post.getViews()+1);
         return post;
+    }
 
+    public boolean isWriterOrAdmin(User user, Post post) {
+        return user.getRole() == Role.ROLE_ADMIN || post.getUser().getId().equals(user.getId());
     }
 }
